requirements.txt
flask
faiss-cpu
sentence-transformers
pypdf2
python-docx
transformers
torch
google-genai

